Change grid stretching parameters, add stretch_uncentered
The new (more uniform) syntax is explained in the par.md documentation.
Essentially, you use e.g. stretch_dim(2) = 'uni' | 'symm' to enable stretching.
Par files have to be adjusted.

stretch_uncentered: if true, the mod_geometry routines try to interpolate
to the cell face (which is not in the middle between stretched cell-centers)