Make m_octree_mg boundary condition code consistent with upstream
The same type of fix (but with slightly different code) was applied in the
octree_mg repository at https://github.com/jannisteunissen/octree-mg and in
the amrvac code files. Now they are consistent again.

I also copied a fix for the 1d case of mg_get_face_coords -- which is
currently not used in amrvac I think